Implementing lease drop action + splitting code
