+++ cvmfs/CMakeLists.txt
@@ -281,6 +281,8 @@ set (CVMFS_SWISSKNIFE_SOURCES
   swissknife_sync.cc swissknife_sync.h
   swissknife_zpipe.cc swissknife_zpipe.h
   swissknife_lease.cc
+  swissknife_lease_curl.cc
+  swissknife_lease_json.cc
   sync_item.cc sync_item.h
   sync_mediator.cc sync_mediator.h
   sync_union.cc sync_union.h
